test: moved and splitted test
